chore: improved index page
