+++ packages/core/src/utils/__tests__/generators.test.mjs
@@ -2,11 +2,70 @@ import assert from 'node:assert/strict';
 import { describe, it } from 'node:test';
 
 import {
+  getEntryDescription,
   groupNodesByModule,
   getVersionFromSemVer,
   coerceSemVer,
   getCompatibleVersions,
 } from '../generators.mjs';
+
+describe('getEntryDescription', () => {
+  it('returns llm_description when available', () => {
+    const entry = {
+      llm_description: 'LLM generated description',
+      content: { children: [] },
+    };
+
+    const result = getEntryDescription(entry);
+    assert.equal(result, 'LLM generated description');
+  });
+
+  it('extracts first paragraph when no llm_description', () => {
+    const entry = {
+      content: {
+        children: [
+          {
+            type: 'paragraph',
+            children: [{ type: 'text', value: 'First paragraph' }],
+          },
+        ],
+      },
+    };
+
+    const result = getEntryDescription(entry);
+    assert.ok(result.length > 0);
+  });
+
+  it('returns empty string when no paragraph found', () => {
+    const entry = {
+      content: {
+        children: [
+          { type: 'heading', children: [{ type: 'text', value: 'Title' }] },
+        ],
+      },
+    };
+
+    const result = getEntryDescription(entry);
+    assert.equal(result, '');
+  });
+
+  it('removes newlines from description', () => {
+    const entry = {
+      content: {
+        children: [
+          {
+            type: 'paragraph',
+            children: [{ type: 'text', value: 'Line 1\nLine 2\r\nLine 3' }],
+          },
+        ],
+      },
+    };
+
+    const result = getEntryDescription(entry);
+    assert.equal(result.includes('\n'), false);
+    assert.equal(result.includes('\r'), false);
+  });
+});
 
 describe('groupNodesByModule', () => {
   it('groups nodes by api property', () => {
